tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.create2.tiny.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.create2.tiny</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.create2.tiny</h1>
<p class="desc">createImageData(sw, sh) works for sizes smaller than one pixel</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("createImageData(sw, sh) works for sizes smaller than one pixel");
_addTest(function(canvas, ctx) {

var imgdata = ctx.createImageData(0.0001, 0.0001);
_assertSame(imgdata.data.length, imgdata.width*imgdata.height*4, "imgdata.data.length", "imgdata.width*imgdata.height*4");
_assertSame(imgdata.width, 1, "imgdata.width", "1");
_assertSame(imgdata.height, 1, "imgdata.height", "1");
var isTransparentBlack = true;
for (var i = 0; i < imgdata.data.length; ++i)
    if (imgdata.data[i] !== 0)
        isTransparentBlack = false;
_assert(isTransparentBlack, "isTransparentBlack");


});
</script>
